refactor(ruborist): extract node-type propagation into resolver/node_types.rs
The prod/dev/optional/peer propagation pass (NodeFlags,
update_node_type_from_edge, compute_node_types and its five unit
tests) is a self-contained graph analysis with no coupling to
resolution; builder.rs had become the catch-all for it. Pure move —
builder re-exports both functions so existing paths keep working.
