+++ crates/ruborist/src/resolver/mod.rs
@@ -8,6 +8,7 @@ pub mod edges;
 pub mod git;
 #[cfg(feature = "http-tarball")]
 pub mod http;
+pub mod node_types;
 pub mod registry;
 pub mod runtime;
 pub mod semver;
